make pixels bigger for gap-hack
If we make the pixels by gap-hack kern bigger in height and width, and
reduce the width of the right-most and the height of the lower most
pixels by the gap-hack kern, we get the correct size of pixels for all
but pixels next to skips.

If the gap-hack kern isn't too big this might give better results than
the implementation of gap-hack with just kerning, and the size of pixels
would be correct. But the parsing is more complicated. I'm not sure
whether this is worth it. Also pixels might "bleed-out" this way (if
gap-hack is set too big).