increase pixel size by gap-hack kern
This way the error introduced by gap-hack regarding the pixel size is
smaller (only the bottom row and the righter most pixels in each row
have the wrong size, as well as those left and above of skips). This way
the error in total width and height of the picture is exactly 1 gap-hack
kern, instead of (wd-1) * gap-hack kern and (ht-1) * gap-hack kern.